+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- CPX-AP: Unittests for cpx_ap
 - CPX-AP: Added INT16 and UINT16 support for MPA modules in read/write functions, expanded system tests
 
 ## v0.4.2 - 29.05.24